Added NN main structure notebook
Includes setup for image upload, preprocessing, conversion to tensors, neural network model definition, training loop (with validation), neural network testing, Of course I haven't trained it on the proper data yet and the model is very basic at the moment and likely won't be effective at its current state. In my opinion however it is a decent starting block for working on building a good model.